将项目更新为springboot项目

+++ src/main/java/cn/edu/guet/controller/TeacherInfoChangeController.java
@@ -6,95 +6,73 @@ import cn.edu.guet.model.Teacher;
 import cn.edu.guet.service.ITeacherInfoChangeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@Controller
+@RestController
 @RequestMapping("TeacherInfoChange")
 public class TeacherInfoChangeController {
 
     @Autowired
     ITeacherInfoChangeService teacherInfoChangeService;
 
-    @Autowired
-    ResponseTemplate responseTemplate;
-
     @RequestMapping(value = "getAllTeachers",method = RequestMethod.GET)
     @ResponseBody
     public ResponseTemplate getAllTeacher() {
         List<Teacher> allTeachers= teacherInfoChangeService.getAllTeachers();
-        responseTemplate.setCode(200);
-        responseTemplate.setMessage("查询成功");
-        responseTemplate.setData(allTeachers);
-        return responseTemplate;
+        return ResponseTemplate.result(200,"查询成功",allTeachers);
     }
 
     @RequestMapping(value = "getTeacherById",method = RequestMethod.GET)
     @ResponseBody
     public ResponseTemplate getTeacherById(String id){
 
-        Teacher teacher = teacherInfoChangeService.getTeacherBydId(id);
-        responseTemplate.setCode(200);
-        responseTemplate.setMessage("查询成功");
-        responseTemplate.setData(teacher);
-        return responseTemplate;
+        if(id !=null){
+            Teacher teacher = teacherInfoChangeService.getTeacherBydId(id);
+            return ResponseTemplate.result(200,"查询成功",teacher);
+        }
+        return ResponseTemplate.result(200,"查询成功",null);
 
     }
 
     @RequestMapping(value = "teacherJobChange",method = RequestMethod.POST)
-    @ResponseBody
     public ResponseTemplate teacherJobChange(@RequestBody Teacher teacher){
-
         try {
-            System.out.println("teacher.getId() = " + teacher.getId());
             teacherInfoChangeService.teacherJobChange(teacher);
-            responseTemplate.setCode(200);
-            responseTemplate.setMessage("更新成功");
-            responseTemplate.setData(teacher);
+            return ResponseTemplate.result(200,"更新成功",teacher);
         } catch (Exception e) {
             e.printStackTrace();
-            responseTemplate.setCode(401);
-            responseTemplate.setMessage("更新失败");
-            responseTemplate.setData(teacher);
+            return ResponseTemplate.result(401,"更新失败",teacher);
         }
-        return responseTemplate;
-
     }
 
     @RequestMapping(value = "deleteTeacher",method = RequestMethod.DELETE)
-    @ResponseBody
     public ResponseTemplate deleteTeacher(String id){
         try {
-            teacherInfoChangeService.deleteTeacher(id);
-
+            if(id !=null){
+                teacherInfoChangeService.deleteTeacher(id);
+                return ResponseTemplate.result(200,"删除成功",null);
+            }
         } catch (Exception e) {
             e.printStackTrace();
+            return ResponseTemplate.result(401,"删除失败",null);
         }
-
-        return null;
-
+        return  ResponseTemplate.result(401,"参数为空",null);
     }
 
     @RequestMapping(value = "addTeacher",method = RequestMethod.POST)
-    @ResponseBody
     public ResponseTemplate addTeacher(@RequestBody Teacher teacher){
 
         try {
-            teacherInfoChangeService.addTeacher(teacher);
-            responseTemplate.setCode(200);
-            responseTemplate.setMessage("添加用户信息成功");
-            responseTemplate.setData(null);
+            if(teacher!=null){
+                teacherInfoChangeService.addTeacher(teacher);
+                return ResponseTemplate.result(200,"添加用户信息成功",null);
+            }
         } catch (Exception e) {
             e.printStackTrace();
-            responseTemplate.setCode(401);
-            responseTemplate.setMessage("添加用户失败");
-            responseTemplate.setData(teacher);
+            return ResponseTemplate.result(401,"添加用户失败",teacher);
         }
-        return responseTemplate;
-
+        return  ResponseTemplate.result(401,"参数为空",null);
     }
 }
